FaceItem/requirements.txt
numpy==1.14.2
opencv-contrib-python==3.4.0.12
pillow
google-cloud-vision
GOOGLE_APPLICATION_CREDENTIALS
